requirements.txt
alive_progress==3.3.0
matplotlib==3.10.6
numpy==2.2.6
pandas==2.3.2
pillow==11.3.0
scipy==1.16.2
scikit-image==0.25.2
scikit-learn==1.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9